requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Data Processing
pandas==2.1.4
numpy==1.26.3
scipy==1.11.4

# Visualization
matplotlib==3.8.2
seaborn==0.13.1
plotly==5.18.0
kaleido==0.2.1

# AI/LLM Integration (Google Gemini - FREE)
google-generativeai==0.8.6
anthropic==0.8.1
langchain==0.1.0
langchain-community==0.0.10
tiktoken==0.5.2

# Background Jobs
celery[redis]==5.3.4
redis<5.0.0,>=4.5.2
flower==2.0.1

# Database & Caching
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
aiosqlite==0.19.0

# HTTP & Async
httpx==0.26.0
aiofiles==23.2.1
aioredis==2.0.1

# Monitoring & Logging
python-json-logger==2.0.7
sentry-sdk==1.39.2
prometheus-client==0.19.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
pyyaml==6.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.26.0
faker==22.0.0

# Code Quality
black==23.12.1
flake8==7.0.0
isort==5.13.2
mypy==1.8.0
pylint==3.0.3

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.3
